Fix what is I think an off-by-one in query and sequence searching.
When taking the mean of the N minimum distances, only consider those 
distances which are closer than the infinite/uninitialized distance.  

(This renders us potentially vulnerable to an internal consistency 
horror where we end up trying to divide by zero.  This is unlikely to be 
picked up by unit tests, but might well be by random tests if this is in 
fact a problem.)


git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@69 7f337319-b335-0410-8187-a933a6394831