Temporary signal handling.
"Temporary" because it will not work on platforms which don't have real 
signals (Hello, Win32) but nevertheless useful, as otherwise gcov (the 
code coverage tool) will not count executions of processes that are 
killed by signals (in particular, the server processes started in the 
tests).

A brief experiment suggests that we're up to about 80% statement 
coverage from our 20 tests.


git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@120 7f337319-b335-0410-8187-a933a6394831