Make a start on refactoring: work the DB header initialization into its own function, used from both initTables and batchinsert.
This actually removes a mmap()/munmap() pair from inside the batchinsert 
loop over files.  Not really sure why that was there in the first place; 
maybe to sync the header to disk more forcefully?  In any case, no 
change in behaviour should happen as a result of this code change.


git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@140 7f337319-b335-0410-8187-a933a6394831