Implement O((M+N) log(M+N)) duplicate key detection in batchinsert, rather than naive O(N^2).
Note that I haven't measured the constants in those complexity 
expressions; I am anticipating that 40000 is a sufficiently large N for 
this to matter.

Write a test case for duplicate keys, too.  Use 0037, since no-one else 
seems to be writing tests, and everything is merged onto the trunk these 
days anyway.


git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@500 7f337319-b335-0410-8187-a933a6394831