Ugly hack upon ugly hack:
* extend adb__query() to take a bunch of other arguments that allow 
  useful functionality through to the SOAP server;
* alter the RadiusQuery reporter so that the count for tracks is 
  returned through SOAP, punning one of the result fields for the 
  purpose;
* alter argv construction to be more dynamic, to reflect that the very 
  presence of some arguments changes audioDB behaviour.

Now test 0020 passes, only some 12 months after it was originally 
written, yay.  Alter test 0050 also, so that the results are actually 
vaguely what would be expected from a radius search; they are not 
completely checked for correctness, but are (apart from the last two) 
the same as test 0040.

I believe that the modifications are backward compatible; lightly 
testing with an old audioDB binary suggests that old-format SOAP queries 
continue to work.  Currently too baby-encumbered to think of how to test 
this fully.


git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@537 7f337319-b335-0410-8187-a933a6394831