Add library SONAME versioning.
This has the unfortunate effect of altering how the linker and the
runtime library resolver find relevant files, which has effects on how
the library test suite should be run.  So...

... also rework how to run libtests.  Start by deleting 28
almost-identical copies of Makefile and run-test.sh, and also the
completely useless copy of test-utils.sh.  Then:

* library tests assume the existence of ../libtest.mk, a file with make
  syntax;
* ../libtest.mk is responsible for providing enough symbolic links to
  con the linker and the runtime resolver into finding our library;
* the default way of doing that is by using the -rpath linker flag.
* run-tests.sh converts from test1 success convention ("return 0") to
  tests success convention ("exit 104")
* clean.sh cleans up our symbolic links.

This test regime stands a reasonable chance of running on OS X
eventually, and a snowball's chance in heaven (remember, heaven is
hotter than hell) on Windows.  It should still be straightforward to
merge libtests/ into tests/ when that is appropriate.

Don't forget to add ../libtest.mk

git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@646 7f337319-b335-0410-8187-a933a6394831