API const correctness.
The char *path arguments to audiodb_open() and audiodb_create() should
be const; make it so.

Also arrange for the datasize, ntracks and datadim arguments to
audiodb_create() to be unsigned to match the fields in the internal
audioDB class.  Technically this is an ABI change, but since nothing is
calling this function with anything other than zero arguments yet
(correct me if I'm wrong) no-one should notice.  (If you notice, shout)

git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@654 7f337319-b335-0410-8187-a933a6394831