Sampling enhancements
Allow specification of a key with -k for doing sampling between a single
datum and the rest of the database.  (The implementation is slightly
icky, as mentioned in the previous commit message; hopefully this
situation will not remain for too long).

While we're at it, allow -r/--resultlength as a synonym for --nsamples
(if --nsamples is not given) in order better to reflect the analogy
between querying and sampling.

git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@972 7f337319-b335-0410-8187-a933a6394831