begin work on sampling API
This is motivated by the need to be able to sample with arbitrary feature data
(e.g. from a feature file) against a database, for the JNMR "collections" paper
revisions or possible ISMIR paper revisions.  That bit doesn't work yet, but
the C-ified version of the current functionality (sample db x db and
sample key x db) works to the level of anecdotal tests.

The general approach is to mirror the _query_spec() API, where a whole heap
of knobs and twiddles are available to the user.  Unlike in the _query_spec()
API, not quite all of the knobs make sense (and even fewer are actually
implemented), but the basic idea is the same.

I pity the poor chump who will have to document all this.

git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@1009 7f337319-b335-0410-8187-a933a6394831